feat(venue): remove VenueMessageingApps old
