+++ src/features/venue/components/VenueMessagingApps/VenueMessagingApps.tsx
@@ -1,23 +1,24 @@
 import React, { useCallback } from 'react'
 import { Social } from 'react-native-share'
 
-import { VenueResponse } from 'api/gen'
-import { MessagingApps } from 'features/share/components/MessagingApps/MessagingApps'
+import { MessagingApps } from 'features/share/components/MessagingAppsNew/MessagingAppsNew'
 import { getShareVenue } from 'features/share/helpers/getShareVenue'
+import { useVenue } from 'features/venue/api/useVenue'
 import { analytics } from 'libs/analytics'
 
 type MessagingAppsProps = {
-  venue: VenueResponse
+  venueId: number
 }
 
-export const VenueMessagingApps = ({ venue }: MessagingAppsProps) => {
-  const { share, shareContent } = getShareVenue({ venue, utmMedium: 'social_media' })
+export const VenueMessagingApps = ({ venueId }: MessagingAppsProps) => {
+  const { data: venue } = useVenue(venueId)
+  const { share, shareContent } = getShareVenue({ venue: venue, utmMedium: 'social_media' })
 
   const messagingAppAnalytics = useCallback(
     (social: Social | 'Other') => {
-      analytics.logShare({ type: 'Venue', venueId: venue.id, from: 'venue', social })
+      analytics.logShare({ type: 'Venue', venueId, from: 'venue', social })
     },
-    [venue.id]
+    [venueId]
   )
 
   if (!shareContent?.url) return null
